+++ CHANGELOG.md
@@ -2,6 +2,17 @@
 
 # Changelog
 
+## 0.3.1
+
+-   Minor bug fixes and code style improvements
+
+## 0.3.0
+
+-   Add cookies to support re-loading state when you come back to the website.
+    -   Light/Dark theme
+    -   Shown/hidden columns
+    -   Column arrangement
+
 ## 0.2.2
 
 -   Add spell name to modal title when viewing long text content